Added temperatures to status
The temperatures come from a different part of the API to the print status so I've moved the API requests into functions which dump the data into global variables to be sent to the app by the fetchall function.
Again, I'm not sure I'm doing it right so beware if you use this code.
The time remaining function needs tweaking to work for times under an hour.